Made game board hold values by storing it as store.gameBoard in app.js
Created a working winner for X using an if statement (&& and ||)

Created winning combinations for O. Need to check to see if it works.

Created an else statement for 'Draw'. Need to check if that works.

Next step is to make a user message for X wins, O wins, and draws

After that, I need to successfully clear the board to start a new game.